Making of Counter in Android
CountDownTimer is a abstract class so it does not has a body.  We can only use it by applying it in other places. So we use object keyword when we will define it. Other two functions below come with this class.  First one, onTick , ask us two things one of them asks what is the number that will be started from to countdown. other ask how much time it should come back.  Second one asks what will it does when it finished.